ttlser vesion bump to 1.0.3 maybe deterministic now?
A reminder to those of you following along at home that if you want
to sort records with more than one element deterministically you
probably shouldn't add the ranks of the elements of the records
together otherwise records composed of elements that are close
together will have a nasty habbit of adding up to the same rank.

Don't add ranks together for sorting!