ontload update and rename template files
ontload fixed a nasty bug where the SciGraph repo would be returned to
the old branch prior to the graphload compelting. Some reafactoring and
moved the services related arguments out. The will reappear with
scigraph-deploy.

Moved graphload and services templates to *.template to be consistent
with upcoming changes for scigraph-deploy.