pax-4th-ed-indexes.txt added structure name
'central amygdaloid nucleus' for the abbreviation Ce.
This structure name was missing entirely from the source,
but is present in the 6th edition. The spatial use of the label
is consistent between versions. This may simply have been an
oversight in the prepration of the indexes of the 4th edition.