core multiple roots for LabelsBase and type checking
Ont now knows how to handle cases where there are multiple roots in a
single file.

Fail loudly when a Source has not been initialized since it can lead
to very cryptic sorting errors down the line.

First shot at type checking arguments that are passed into a class
may also need it for ontologies. It is not currently enabled because
it does not help at the moment and I don't think it is really the
right way to do this.