parc imported and local LabelsBase now play nice
Turns out calls to __subclasses__ do not behave consistently and will
variously return '__main__.Class' or 'module.Class' depending on
whether the superclass is in the same file where __subclasses__ is
called or not. This means that to have everything play nice we have to
look to see wehther a class has both a local copy in globals as well
as a copy in the current module.