added methods_core.py and methods.py
This covers the perliminary work in developing the core schema and the
rough outline for many of the classes that we will need to model. Only a
tiny subset of the methods are correctly modelled at the moment.

Also introduces some of the new infrastructure for writing owl in python
without using rdflib.infix owl.