added resources WHSSD v1 labels file
This is an older version of the file.

Based on the file creation time metadata that I have for this file
`5.5K Aug 31  2016 WHS_SD_rat_atlas_v1.label` matches the date of
83044d8f032331b01a5442a70011e81a7c3dbfda it is possible that i retrieved
it from the same source, but I do not have a record of using the wget
method described, so I may have downloaded it using a browser but still
from the same source.