parc updated and test only tracked *py files
Broke the waxholm space parcellation generation code into its own file
for easier maintenance.

Updated FreeSurferCL to have versions to mirror the approach we have
taken with paxinos. It may be more verbose in terms of annotations but
that way all of the variants can be preserved over time. It may be worth
looking into a set of rules for deduping prov on names that doe not
change, ie that things without prov are present in all versions without
changes.