resolver version-lookup.lua updated
The github api limits the amount of time any query can take. The <= on
date ranges was taking way too long and returning incomplete results,
now we just search the previous 28 days to keep the query time down so
that we get complete results that can actually be ordered correctly.