ttlser resolved nosort issue + tests
Even though we were not sorting the visible elements of ordered lists we
were still sorting their ranks, which meant that things with ranks that
were symmetric, such as 7, 2 and 2, 7 would appear to randomly switch
places. ListRanker now adds nodes where reorder = False to the nosort
set so that we can skip them.

Also reworked test_ttlser to be much more modular and easier to extend
for any new formats.