parcArts literally had to reimport a class defined in the file
thank's python!

Remember kids, when you move your modules around, if there is even a
tiny chance that you will be accessing a module defined in the the
via import by some other name, you are going to have a very very bad
time, because __main__.Class and __init__.Class are not actually the
same class as package.module.Class. Have fun!