pyontutils core OntRes for managing ontology resources
This object representation is more in line with the ontquery approach,
except that it is much better thought out than the services approach
for services that are derived from serialized sources. These classes
may ultimately be moved to ontquery once I figure out a better way to
decouple the rdflib dependency for use cases that want only the bare
bones ontology id/term representation.

The most useful feature for OntRes is that OntHeader makes it possible
to retrieve only the header portion of a remote ontology file, making
it much faster to extract import chains.