config warn on missing ~/.config folder
when importing pyontutils in a situation where the user has no
~/.config, for example a service only user, we need to warn that the
folder is missing. The alternative is to package things like
curie_map.yaml on a per-release basis, however this means that users
cannot modify curies and they are static per system (this happens with
ontquery right now). We need this warning, but it is a temporary
solution for the larger question of how to manage the default curies for
a release while providing a way for users to configure their own curies
in a way that doesn't do surprising things like go to the network for
for data like we do right now.