Made some edits
I think that the material I prepared for the slide set makes a better introduction than what we have.  Also, why are we giving term lists in multiple places?  We first mention them in the table, then there are several other headings that refer to term lists. I think that's confusing.  They should only be in one place.