neurondm core location model part 3
As it turns out ELK cannot reason about ObjectUnionOf assertions so
the way that we were modelling location phenotypes simply does not
work. The solution is to add an axiom making every region used in
a location phenotype part of itself. This allows us to use a single
partOf restriction on the region itself so that the region is returned
when we query for a part of that region. ELK is happy with this and
it is pretty fast.