scigraph codegen dynamic queries warn on double output error
The issue we were having with the bundles query endpoint was that it
didn't set start-id as a parameter explicitly, so codegen didn't know
anything about it. This maniested as a completely cryptic error about
the dynamic endpoint function receiving two values for the output
argument. Thus we catch type errors and add a slightly more helpful
message.