test ttlser determinism failure found
It was id collisions from the urn all along. Turns out that left
aligning is the same a right padding, which causes numbers like 999 and
999000 to collide because 999 is padded to 999000 not to 000999. So now
that we replace < with > the collisions should stop. Apparently I missed
the additional pattern that all my colliding ids ended with zero, but
here we are at long last.

closes https://github.com/tgbugs/pyontutils/issues/81