methods split TEMP into local and local:blank/ namespaces
This is the first step of the split.
Trying to change both at the same time makes it much harder to isolate
the class identifier resequnecing because the contents of the local
subgraphs also change. This way the sequential changes will make it
simple to derive the mapping between the old TEMP and the new local
ids without having to store the mapping explicitly.

I think that this convention is better than trying to track a global
ilxtib sequence for now, at least as long as there is a 1:1 mapping
between indexes and files since we can read the index back in from the
serialized file if needsbe, more work would have to be done on this if
we really want it to work, next up is testing to see whether we can
recover the TEMP -> local mapping from the diffs, note that this
splitting of sequences is ok because total ordering is preserved since
we are just removing the gaps in the two sequences

The new convention for how to handle this is as follows.

:ontology a owl:Ontology ;
ilxtr:indexNamespace local: .

local:blank/ by convention single use ids to differentiate equivalent
classes that have structure that is otherwise the same

local:{index-name}/ could be used for any other file local sequences
without having to track things globally but does require 1:1 mapping
between files and ontologies during development