fix IdentityBNode.__hash__, OntGraph.subjectIdentity
working on OntGraph.subjectsRenamed for map-identifiers uncovered a
number of issues with subjectIdentity (subject was not factored out)
and IdentityBNode, terrifyingly the ibnode bug is that str.__hash__
was resulting in two idbnodes with the same identity not hashing to
the same value

methods_mapping is a prototype for being able to simply mark a commit
as an id renaming operations and not having to store the mapping
explicitly, this is the dual of the previously implemented approach
where the remapping is defined explicitly and the transformation is
stored as the renames. This new version is a better fit for working
with data from git and the previous version can be used to generate
the new graph for commiting so that git can store the diff and we can
recover the changes using the new way.