start of work on IdentityPython for ontload
use case for this is to deterministically identify the contents of yaml
files e.g. those that are used for scigraph graphload config

ontload start of rework of the process for graphload so that the base of
the config file can be included in the name of working folder so that we
can load multiple different configurations from the same repo (e.g. for
sparc vs scicrunch ontology loads) as part of our CI workflows