scigraph readme implement scigraph-load-graphs
A vastly simplified cli interface that takes a list of input paths or
iris and creates ontologies.yaml for you, elk is run on all inputs and
you can specify the name of the output zip, if you are in a real pinch
you can also point to a different graphload template yaml file. I get
the sense that this is much closer to what I have wanted in terms of
composability and we save the actual graphload template that was used
in the built graph for the record anyway. Ironically if I knew java I
probably would have just written it to be part of scigraph itself
without the need for all the extra templates and configs and such.

This is one step closer to being able to get rid of ontload entirely
and replace it with just the (very fast) part that can be used to
rewrite the import chain to point to local files as well as fetch
external ontologies in a more efficient manner.