more shuffling to make it possible to reduce startup times
e.g. for scig which is now down to 200ms instead of 2.5 seconds or
some utter nonsense

asyncio taks a couple hundred milliseconds, so if you don't need it,
the overhead is annoyingly large, there are numerous other examples

the changes should be invisible externally, but code that imports from
pyontutils should now be able to reduce startup times by tweaking the
exact imports e.g. to use utils_fast