vbump for release, drop 3.6 support
this is the first release after the merge of the ibnode-v3 work
ibnode behavior has been regularized and validated using 4ish different
implementations, two in python, one in racket, and for fully named
triples (no bnodes) one in postgres plpgsql, a full explication of the
spec is forthcoming, the short version is contained in the enum types,
and since there are multiple ways that one might id a graph, ibnode
needs to support all of them, and more or less does

one area that need improvement is caching, at the moment, trying to
process large graphs will lead to ever-increasing memory usage if the
cache is not periodically reset, e.g. when switching between subjects,
some optimization has been made so that e.g. predicates are retained

one area where I'm not sure about the stability and correctness of the
implementation is for graphs with cycles, extensive work has been put in
to make it possible to get identities for graphs with cycles, and a
first pass at de-cycling to get consistent identities is done, however
I'm fairly certain that the current implementation will not behavior
correctly in certain slightly asymmetric cases where, e.g. otherwise
seemingly identical nodes in a cycle differ because of what they are
connected to outside the cycle

in short, if your graph has cycles you can get an identity for it, but
it probably isn't a stable or reliable identity, so I strongly reccomend
not using this version to identify graphs with cycles, for practical use
this is fine anyway because a graph with cycles found in the wild is
probably evil and you don't want to ingest it until it is fixed upstream

updated all the setup.py files with current classifiers and dropped
support for 3.6, these may still work on 3.6 but since there is no
longer an easy way for me to test and 3.6 was eol nearly 4 years ago,
it is time to move on

update MANIFEST.in to include evil.ttl for testing

skip test_subject_identities, too slow for normal use

bump augpathlib and idlib deps