ttlser fix rank_vec index error, vbump
There are cases where multiple types may be present in a graph that
are subClassOf URIRef (e.g. ontquery OntId). If that happens and those
types are used as predicates along with URIRef then the number of
distinct predicates in the graph could be larger than the number of
unique strings for predicates in the graph. As a result, trying to use
in ser.predicateOrder to partition graph.predicates by iterating over
ser.predicateOrder and the supposed complement in graph.predicates
would leave out any duplicates that were str.__eq__ to eachother but
that were not in ser.predicateOrder, this caused an IndexError in
CustomTurtleSerializer._BNodeRank.specref. This is now fixed and a
test has been added that can catch the issue if it happens again.

Note that it is bad to have anything other than type(_) is URIRef in
the graph but this is not the right place to catch that error, and we
want to be able to debug output of such potentially malformed graphs
because the output often gives a hit about what is broken.